TG-46:delete error fixed, button back added
