+++ src/Components/Employees/EmployeeForm.js
@@ -1,6 +1,7 @@
 import React from 'react';
 import styles from './employees.module.css';
 import { useEffect, useState } from 'react';
+import Button from './Button';
 
 const path = window.location.pathname.split('/');
 let employeeId = path[path.length - 1];
@@ -62,68 +63,73 @@ const EmployeeForm = () => {
   };
 
   return (
-    <form className={styles.addform} onSubmit={onSubmit}>
-      <div className={styles.formcontrol}>
-        <label>Name</label>
-        <input
-          name="name"
-          type="text"
-          placeholder="name"
-          value={userInput.name}
-          onChange={handleChange}
-        />
-      </div>
-      <div className={styles.formcontrol}>
-        <label>Last Name</label>
-        <input
-          name="lastName"
-          type="text"
-          placeholder="last name"
-          value={userInput.lastName}
-          onChange={handleChange}
-        />
-      </div>
-      <div className={styles.formcontrol}>
-        <label>Email</label>
-        <input
-          name="email"
-          type="mail"
-          placeholder="email"
-          value={userInput.email}
-          onChange={handleChange}
-        />
-      </div>
-      <div className={styles.formcontrol}>
-        <label>Phone Number</label>
-        <input
-          name="phone"
-          type="number"
-          placeholder="phone"
-          value={userInput.phone}
-          onChange={handleChange}
-        />
-      </div>
-      <div className={styles.formcontrol}>
-        <label>Password</label>
-        <input
-          name="password"
-          type="password"
-          placeholder="password"
-          value={userInput.password}
-          onChange={handleChange}
-        />
-      </div>
-      <button
-        type="submit"
-        value="Save Employee"
-        className={styles.btn}
-        onClick={(e) => {
-          onSubmit(e);
-        }}
-      >
-        Confirm
-      </button>
-    </form>
+    <>
+      <form className={styles.addform} onSubmit={onSubmit}>
+        <div className={styles.formcontrol}>
+          <label>Name</label>
+          <input
+            name="name"
+            type="text"
+            placeholder="name"
+            value={userInput.name}
+            onChange={handleChange}
+          />
+        </div>
+        <div className={styles.formcontrol}>
+          <label>Last Name</label>
+          <input
+            name="lastName"
+            type="text"
+            placeholder="last name"
+            value={userInput.lastName}
+            onChange={handleChange}
+          />
+        </div>
+        <div className={styles.formcontrol}>
+          <label>Email</label>
+          <input
+            name="email"
+            type="mail"
+            placeholder="email"
+            value={userInput.email}
+            onChange={handleChange}
+          />
+        </div>
+        <div className={styles.formcontrol}>
+          <label>Phone Number</label>
+          <input
+            name="phone"
+            type="number"
+            placeholder="phone"
+            value={userInput.phone}
+            onChange={handleChange}
+          />
+        </div>
+        <div className={styles.formcontrol}>
+          <label>Password</label>
+          <input
+            name="password"
+            type="password"
+            placeholder="password"
+            value={userInput.password}
+            onChange={handleChange}
+          />
+        </div>
+        <div className={styles.divBtn}>
+          <button
+            type="submit"
+            value="Save Employee"
+            className={styles.btn}
+            onClick={(e) => {
+              onSubmit(e);
+            }}
+          >
+            Confirm
+          </button>
+        </div>
+      </form>
+      <Button color="black" text={'back'} href={'/employees'} />
+    </>
   );
 };
